+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Fix possible bug with grand arena enemy selection and/or logging
 * Use last `N` battles for training
+* Update `User-Agent`
 
 # `2.2`