+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# `latest`
+# `2.3`
 
 * Fix possible bug with grand arena enemy selection and/or logging
 * Use last `N` battles for training